requirements.txt
python-pptx
pandas
numpy
pretty_html_table
nltk

